requirements.txt
django
mysql-connector-python
djangorestframework
